% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cja_get_audit_logs_search.R
\name{cja_get_audit_logs_search}
\alias{cja_get_audit_logs_search}
\title{CJA Get Audit Logs Search}
\usage{
cja_get_audit_logs_search(
  client_id = Sys.getenv("CJA_CLIENT_ID"),
  client_secret = Sys.getenv("CJA_CLIENT_SECRET"),
  org_id = Sys.getenv("CJA_ORGANIZATION_ID"),
  debug = FALSE
)
}
\arguments{
\item{client_id}{Set in environment args, or pass directly in the function}

\item{client_secret}{Set in environment args, or pass directly in the function}

\item{org_id}{Set in environment args or pass directly in the function}

\item{debug}{Set to TRUE if needed to help troubleshoot api call errors}
}
\value{
A data frame of audit logs and corresponding metadata
}
\description{
This function will pull a list of audit logs.
}
\examples{
\dontrun{
get_me()
}
}
