% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cja_get_dimensions.R
\name{cja_get_dimensions}
\alias{cja_get_dimensions}
\title{Get a list of dimensions in CJA}
\usage{
cja_get_dimensions(
  dataviewId = NULL,
  expansion = "description",
  includeType = NULL,
  locale = "en_US",
  debug = FALSE,
  client_id = Sys.getenv("CJA_CLIENT_ID"),
  client_secret = Sys.getenv("CJA_CLIENT_SECRET"),
  org_id = Sys.getenv("CJA_ORGANIZATION_ID")
)
}
\arguments{
\item{dataviewId}{\emph{Required} The id of the dataview for which to retrieve Dimensions}

\item{expansion}{Comma-delimited list of additional segment metadata fields to include on response. See Details for all options available.}

\item{includeType}{Include additional segments not owned by user. Options include: "shared" "templates" "deleted" "internal"}

\item{locale}{Locale - Default: "en_US"}

\item{debug}{Used to help troubleshoot api call issues. Shows the call and result in the console}

\item{client_id}{Set in the global environment using "AW_CLIENT_ID" or pass directly here}

\item{client_secret}{Set in the global environment using "AW_CLIENT_SECRET" or pass directly here}

\item{org_id}{Set in the global environment using "AW_ORGANIZATION_ID" or pass directly here}
}
\value{
A data frame of dimensions in a specified dataview
}
\description{
Retrieves a list of dimensions available in a specified dataview
}
\details{
\emph{Expansion} options can include the following:
"approved" "favorite" "tags" "usageSummary" "usageSummaryWithRelevancyScore"
"description" "sourceFieldId" "segmentable" "required" "hideFromReporting"
"hidden" "includeExcludeSetting" "fieldDefinition" "bucketingSetting"
"noValueOptionsSetting" "defaultDimensionSort" "persistenceSetting"
"storageId" "tableName" "dataSetIds" "dataSetType" "type" "schemaPath"
"hasData" "sourceFieldName" "schemaType" "sourceFieldType" "fromGlobalLookup"
"multiValued" "precision"
}
\examples{
\dontrun{
cja_get_dimensions(dataviewId = "dv_5f4f1e2572ea0000003ce262")
}
}
