% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cja_get_me.R
\name{cja_get_me}
\alias{cja_get_me}
\title{CJA Get My Information}
\usage{
cja_get_me(
  expansion = NULL,
  useCache = TRUE,
  debug = FALSE,
  client_id = Sys.getenv("CJA_CLIENT_ID"),
  client_secret = Sys.getenv("CJA_CLIENT_SECRET"),
  org_id = Sys.getenv("CJA_ORGANIZATION_ID")
)
}
\arguments{
\item{expansion}{The endpoint for that particular report}

\item{useCache}{cache the response for faster recall}

\item{debug}{Used to help troubleshoot api call issues. Shows the call and result in the console}

\item{client_id}{Set in environment args, or pass directly here}

\item{client_secret}{Set in environment args, or pass directly here}

\item{org_id}{The organization ID the data should be pulled from}
}
\value{
A list of the current user metadata
}
\description{
This function will quickly pull the list of company ids that you have access to.
}
\examples{
\dontrun{
cja_get_me()
}
}
