% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cja_get_metrics.R
\name{cja_get_metrics}
\alias{cja_get_metrics}
\title{Get a list of metrics in CJA}
\usage{
cja_get_metrics(
  dataviewid = NULL,
  expansion = "definition",
  includeType = NULL,
  locale = "en_US",
  debug = FALSE,
  client_id = Sys.getenv("CJA_CLIENT_ID"),
  client_secret = Sys.getenv("CJA_CLIENT_SECRET"),
  org_id = Sys.getenv("CJA_ORGANIZATION_ID")
)
}
\arguments{
\item{dataviewid}{\emph{Required} The id of the dataview for which to retrieve metrics}

\item{expansion}{Comma-delimited list of additional segment metadata fields to include on response. See Details for all options available.}

\item{includeType}{Include additional segments not owned by user. Options include: "shared" "templates" "deleted" "internal"}

\item{locale}{Locale - Default: "en_US"}

\item{debug}{Used to help troubleshoot api call issues. Shows the call and result in the console}

\item{client_id}{Set in the global environment using "AW_CLIENT_ID" or pass directly here}

\item{client_secret}{Set in the global environment using "AW_CLIENT_SECRET" or pass directly here}

\item{org_id}{Set in the global environment using "AW_ORGANIZATION_ID" or pass directly here}
}
\value{
A data frame of metrics in a specified dataview
}
\description{
Retrieves a list of metrics available in a specified dataview
}
\details{
\emph{Expansion} options can include the following:
"approved" "favorite" "tags" "usageSummary" "usageSummaryWithRelevancyScore" "description" "sourceFieldId" "segmentable" "required" "hideFromReporting" "hidden" "includeExcludeSetting" "fieldDefinition" "bucketingSetting" "noValueOptionsSetting" "defaultmetricsort" "persistenceSetting" "storageId" "tableName" "dataSetIds" "dataSetType" "type" "schemaPath" "hasData" "sourceFieldName" "schemaType" "sourceFieldType" "fromGlobalLookup" "multiValued" "precision"
}
\examples{
\dontrun{
cja_get_metrics()
}
}
