% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cja_top_items.R
\name{cja_top_items}
\alias{cja_top_items}
\title{Get a list of dimensions in CJA}
\usage{
cja_top_items(
  dataId = NULL,
  dimension = "page",
  locale = "en_US",
  dateRange = NULL,
  search_clause = NULL,
  startDate = NULL,
  endDate = NULL,
  searchAnd = NULL,
  searchOr = NULL,
  searchNot = NULL,
  searchPhrase = NULL,
  includeOberonXml = NULL,
  lookupNoneValues = TRUE,
  limit = 10,
  page = 0,
  debug = FALSE
)
}
\arguments{
\item{dataId}{Data Group or Data View to run the report against. Required}

\item{dimension}{Dimension to run the report against. Example: "page" Required}

\item{locale}{Locale Default: "en_US"}

\item{dateRange}{Format: YYYY-MM-DD/YYYY-MM-DD}

\item{search_clause}{General search string; wrap with single quotes. Example: 'PageABC'}

\item{startDate}{Format: YYYY-MM-DD}

\item{endDate}{Format: YYYY-MM-DD}

\item{searchAnd}{Search terms that will be AND-ed together. Space delimited.}

\item{searchOr}{Search terms that will be OR-ed together. Space delimited. See Details}

\item{searchNot}{Search terms that will be treated as NOT including. Space delimited. See Details}

\item{searchPhrase}{A full search phrase that will be searched for. See Details}

\item{includeOberonXml}{Controls if Oberon XML should be returned in the response - DEBUG ONLY. Default: false}

\item{lookupNoneValues}{Controls None values to be included. Default: TRUE}

\item{limit}{Number of results per page. Default: 10}

\item{page}{Page number (base 0 - first page is "0"). Default: 0}

\item{debug}{Use this to see the api calls in the console when trying to debug}
}
\value{
A data frame of dimensions in a specified dataview
}
\description{
Retrieves a list of dimensions available in a specified dataview
}
\details{
**DateRange
Get the top X items (based on paging restriction) for the specified dimension and dataId. Defaults to last 90 days.
\strong{Search Clause examples:} contains test: 'test'
contains test or test1: 'test' OR 'test1'
contains test and test1: 'test' AND 'test1'
contains test and not (test1 or test2): 'test' AND NOT ('test1' OR 'test2')
does not contain test: NOT 'test'
}
\examples{
\dontrun{
cja_get_dimensions(dataviewId = "dv_5f4f1e2572ea0000003ce262")
}
}
