% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_update.R
\name{cja_update_project}
\alias{cja_update_project}
\title{Update a project in CJA}
\usage{
cja_update_project(project = NULL, locale = "en_US", debug = FALSE)
}
\arguments{
\item{project}{The updated/edited project json string}

\item{locale}{Locale Default value : en_US}

\item{debug}{Set to \code{TRUE} to publish the full JSON request(s) being sent to the API to the console when the
function is called. The default is \code{FALSE}.}
}
\value{
A json string
}
\description{
Edits an existing project in CJA
}
