% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_build.R
\name{filter_build}
\alias{filter_build}
\title{Build the Segment}
\usage{
filter_build(
  dataId = NULL,
  name = "this is the name",
  description = "this is the description",
  containers = NULL,
  predicates = NULL,
  context = "hits",
  conjunction = "and",
  exclude = FALSE,
  version = list(1, 0, 0),
  debug = FALSE,
  client_id = Sys.getenv("CJA_CLIENT_ID"),
  client_secret = Sys.getenv("CJA_CLIENT_SECRET"),
  org_id = Sys.getenv("CJA_ORGANIZATION_ID")
)
}
\arguments{
\item{dataId}{This is the data view id that the filter reference.}

\item{name}{This is the name of the new filter}

\item{description}{This is the description of the filter}

\item{containers}{List of the container(s) that make up the filter}

\item{predicates}{List of the predicate(s) to create a filter}

\item{context}{This defines the main container context, either hits, visits (Default), or visitors}

\item{conjunction}{This will tell how the different containers and predicates should be compared. Use either 'and' or 'or' at this time. Sequential 'then' will be added in a future update.}

\item{exclude}{excludes the main container which will include all predicates. Only used when the predicate arguments are used.}

\item{version}{This is the default version. Only used if updating an existing filter. Not to be edited at this time.}

\item{debug}{Helps in troubleshooting function issues}

\item{client_id}{This is the report suite that the filter will be referenced to.}

\item{client_secret}{This is the report suite that the filter will be referenced to.}

\item{org_id}{This is the organization id from the Adobe console project associated with your authentication project.}
}
\description{
This function combines predicates into a container.
}
