% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_container.R
\name{filter_con}
\alias{filter_con}
\title{Create the Segment Container}
\usage{
filter_con(
  context = "hits",
  conjunction = "and",
  predicates = NULL,
  exclude = FALSE
)
}
\arguments{
\item{context}{One of hits, visits, or visitors. Also known as scope}

\item{conjunction}{How should the predicates be combined?  And is the default.}

\item{predicates}{List of predicates created using \code{seg_pred()} function. Must wrapped in a list() function.}

\item{exclude}{Exclude the entire container}
}
\value{
A list object for containers that will be used to build a new filter
}
\description{
This function combines predicates into a container.
}
