% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_predicate.R
\name{filter_pred}
\alias{filter_pred}
\title{Create the Filter Predicate
This function creates the simple predicate of a filter.}
\usage{
filter_pred(
  subject = "page",
  verb = "contains",
  object = "pagename",
  is_distinct = FALSE,
  dataid = NULL
)
}
\arguments{
\item{subject}{This is the subject of the predicate. It is the dimension or metric id.}

\item{verb}{Choose from any of the 30 different verbs}

\item{object}{this is the object of the predicate and answer the question what or how many}

\item{is_distinct}{count the distinct items to compare against instead of dimension number}

\item{dataid}{This is the data view id that the filter will be referencing.}
}
\description{
Create the Filter Predicate
This function creates the simple predicate of a filter.
}
