% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_validate.R
\name{filter_val}
\alias{filter_val}
\title{Validate a filter in CJA}
\usage{
filter_val(filter_body = NULL, debug = FALSE)
}
\arguments{
\item{filter_body}{The json string of the filter that is being validated (required)}

\item{debug}{This enables the api call information to show in the console for
help with debugging issues. default is FALSE}
}
\value{
A validation True or False response
}
\description{
Returns a filter validation for a filter contained in a json string object.
}
