% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cja_get_audit_logs.R
\name{cja_get_audit_logs}
\alias{cja_get_audit_logs}
\title{CJA Get Audit Logs}
\usage{
cja_get_audit_logs(
  client_id = Sys.getenv("CJA_CLIENT_ID"),
  client_secret = Sys.getenv("CJA_CLIENT_SECRET"),
  org_id = Sys.getenv("CJA_ORGANIZATION_ID")
)
}
\arguments{
\item{client_id}{Set in environment args, or pass directly in the function}

\item{client_secret}{Set in environment args, or pass directly in the function}

\item{org_id}{Set in environment args or pass directly in the function}
}
\value{
A data frame of company ids and company names
}
\description{
This function will pull a list of company ids that you have access to.
}
\examples{
\dontrun{
get_me()
}
}
