% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cja_get_audit_logs_search.R
\name{cja_get_audit_logs_search}
\alias{cja_get_audit_logs_search}
\title{Get audit logs search}
\usage{
cja_get_audit_logs_search(body = NULL, debug = FALSE)
}
\arguments{
\item{body}{The json string with the search functions included}

\item{debug}{Set to TRUE if needed to help troubleshoot api call errors}
}
\value{
A data frame of audit logs and corresponding metadata
}
\description{
This function will pull a list of audit logs.
}
\examples{
\dontrun{
cja_get_audit_logs_search(body = jsonrequest)
}
}
