% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cja_get_me.R
\name{cja_get_me}
\alias{cja_get_me}
\title{Get my information}
\usage{
cja_get_me(expansion = NULL, debug = FALSE)
}
\arguments{
\item{expansion}{Comma-delimited list of additional metadata fields to include in the response. Options are 'admin'}

\item{debug}{Used to help troubleshoot api call issues. Shows the call and result in the console}
}
\value{
A list of the current user metadata
}
\description{
This function will quickly pull the list of company ids that you have access to
}
\examples{
\dontrun{
cja_get_me()
}
}
